+++ src/Parser/JsonParser.php
@@ -17,11 +17,6 @@ use function json_decode;
  */
 final class JsonParser implements ParserInterface
 {
-    private bool $convertToAssociativeArray;
-    /** @psalm-var int<1, 2147483647>  */
-    private int $depth;
-    private int $options;
-
     /**
      * @param bool $convertToAssociativeArray Whether objects should be converted to associative array during parsing.
      * @param int $depth Maximum JSON recursion depth.
@@ -30,13 +25,10 @@ final class JsonParser implements ParserInterface
      * @psalm-param int<1, 2147483647> $depth
      */
     public function __construct(
-        bool $convertToAssociativeArray = true,
-        int $depth = 512,
-        int $options = JSON_THROW_ON_ERROR | JSON_INVALID_UTF8_IGNORE
+        private readonly bool $convertToAssociativeArray = true,
+        private readonly int $depth = 512,
+        private readonly int $options = JSON_THROW_ON_ERROR | JSON_INVALID_UTF8_IGNORE,
     ) {
-        $this->convertToAssociativeArray = $convertToAssociativeArray;
-        $this->depth = $depth;
-        $this->options = $options;
     }
 
     public function parse(string $rawBody)